Change training setup

+++ training/run_training.py
@@ -281,7 +281,9 @@ def main():
                 num_nodes=1,              # Set >1 for multi-machine setups
                 # use_distributed_sampler=False
                 accumulate_grad_batches=4,
-                precision="16-mixed"
+                # precision="16-mixed",
+                precision="bf16-mixed",
+                # enable_checkpointing=False
             )
         else:
             print("calling single processing")
